+++ CHANGELOG.md
@@ -6,11 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-
 ## [Unreleased]
 
-## [0.1.0] - 2022-11-21
+### Added
 
+- Add cleaner for Load Balancer Pools.
+- Add cleaner for DNAT rules.
+
+
+## [0.1.0] - 2022-11-21
 
 
 [Unreleased]: https://github.com/giantswarm/cluster-api-cleaner-cloud-director/compare/v0.1.0...HEAD